+++ MLE/MLE/estimatewithcv.py
@@ -12,7 +12,7 @@ def estimatewithcv(datagen, rangefn, rmin=0, rmax=1, raiseonerr=False):
 
     sumwsq = 0
     for c, w, r, cvs in datagen():
-        sumcvsq = np.zeros_like(cvs)
+        sumcvsq = np.zeros_like(cvs, dtype='float64')
         break
 
     n = 0
